backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
motor==3.7.1
pymongo==4.11.3
httpx==0.28.1
pydantic-settings==2.10.1
pillow==11.3.0
numpy==2.1.3
tensorflow==2.19.0
